l10n: Set the description to "System Restore Utility"
The first reason for this change is to make the description in the about dialog
match the GenericName and comment string in the /usr/share/applications/.desktop
file. We want these to match because we'll need that string in the POT file
to have it translated in Launchpad. This will ensure we'll be able to generate
a fully translated .desktop file in the near future.

The second reason for the change is that "for Linux" is redundant and not
relevant to users. Timeshift is a great app name, "System Restore Utility" a great
comment, it's explicit and clear in the app menu, but "for Linux" brings nothing
extra.