Rename desktop file to match binary name
Appsys matches binaries at runtime with desktop files.

The binary name is timeshift-gtk but the desktop file was timeshift.desktop.

This resulted in the window icon being pixelated.

This commit renames the desktop file so that it is found by appsys,
resulting in the application window finding the hicolor icons.

This fixes the pixelation in Alt-tab for Timeshift in environments such as
Cinnamon and GNOME.