Fix missing hourly backups; issues #497 & 467
The code should be changed so that choosing a snapshot to tag follows timing similar to determining if a snapshot for a tag is needed. That is, allow for the possibility that not all runs of cron start right on the same second. A safe change to do this would be to change the timestamp comparison to account for extra seconds.