Update About window
Updated the About dialog to better match the Gtk.AboutDialog

We are not using the Gtk.AboutDialog since clicking any link will launch the web browser with root privileges. Running the web browser as root creates a security risk, and most web browsers will refuse to start when executed as root user.